Adicionado recursos para retornar melhor volta

+++ Q3/corrida/src/main/java/com/vivo/corrida/model/RankingCorridaResponse.java
@@ -15,6 +15,6 @@ public class RankingCorridaResponse {
     private String nomeSuperHeroi;
     private Long quantidadeVoltasCompletas;
     private Long tempoTotalSuperHeroi;
-    private Long melhorVoltaSuperHeroi;
+    private Integer melhorVoltaSuperHeroi;
     private Double velocidadeMediaSuperHeroi;
 }
